Refactor: re-organize revocation.go
